Module google_autoscaler1_beta2::api

Structs

Autoscaler
Cloud Autoscaler resource.

AutoscalerDeleteCall
Deletes the specified Autoscaler resource.

AutoscalerGetCall
Gets the specified Autoscaler resource.

AutoscalerHub
Central instance to access all AutoscalerHub related resource activities

AutoscalerInsertCall
Adds new Autoscaler resource.

AutoscalerListCall
Lists all Autoscaler resources in this zone.

AutoscalerListResponse
There is no detailed description.

AutoscalerMethods
A builder providing access to all methods supported on autoscaler resources. It is not used directly, but through the AutoscalerHub hub.

AutoscalerPatchCall
Update the entire content of the Autoscaler resource. This method supports patch semantics.

AutoscalerUpdateCall
Update the entire content of the Autoscaler resource.

AutoscalingPolicy
Cloud Autoscaler policy.

AutoscalingPolicyCpuUtilization
CPU utilization policy.

AutoscalingPolicyCustomMetricUtilization
Custom utilization metric policy.

AutoscalingPolicyLoadBalancingUtilization
Load balancing utilization policy.

DeprecationStatus
There is no detailed description.

Operation
There is no detailed description.

OperationError
There is no detailed description.

OperationErrorErrors
There is no detailed description.

OperationList
There is no detailed description.

OperationWarnings
There is no detailed description.

OperationWarningsData
There is no detailed description.

Zone
There is no detailed description.

ZoneList
There is no detailed description.

ZoneListCall
A builder for the list method supported by a zone resource. It is not used directly, but through a ZoneMethods instance.

ZoneMaintenanceWindows
There is no detailed description.

ZoneMethods
A builder providing access to all methods supported on zone resources. It is not used directly, but through the AutoscalerHub hub.

ZoneOperationDeleteCall
Deletes the specified zone-specific operation resource.

ZoneOperationGetCall
Retrieves the specified zone-specific operation resource.

ZoneOperationListCall
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethods
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the AutoscalerHub hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.