Struct google_autoscaler1_beta2::AutoscalingPolicy

pub struct AutoscalingPolicy {
    pub custom_metric_utilizations: Option<Vec<AutoscalingPolicyCustomMetricUtilization>>,
    pub max_num_replicas: Option<i32>,
    pub cpu_utilization: Option<AutoscalingPolicyCpuUtilization>,
    pub min_num_replicas: Option<i32>,
    pub cool_down_period_sec: Option<i32>,
    pub load_balancing_utilization: Option<AutoscalingPolicyLoadBalancingUtilization>,
}

Cloud Autoscaler policy.

This type is not used in any activity, and only used as part of another schema.

Fields
custom_metric_utilizations: Option<Vec<AutoscalingPolicyCustomMetricUtilization>>

Configuration parameters of autoscaling based on custom metric.

max_num_replicas: Option<i32>

The maximum number of replicas that the Autoscaler can scale up to.

cpu_utilization: Option<AutoscalingPolicyCpuUtilization>

Exactly one utilization policy should be provided. Configuration parameters of CPU based autoscaling policy.

min_num_replicas: Option<i32>

The minimum number of replicas that the Autoscaler can scale down to.

cool_down_period_sec: Option<i32>

The number of seconds that the Autoscaler should wait between two succeeding changes to the number of virtual machines. You should define an interval that is at least as long as the initialization time of a virtual machine and the time it may take for replica pool to create the virtual machine. The default is 60 seconds.

load_balancing_utilization: Option<AutoscalingPolicyLoadBalancingUtilization>

Configuration parameters of autoscaling based on load balancer.