Struct google_autoscaler1_beta2::ZoneOperationMethods

pub struct ZoneOperationMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the AutoscalerHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_autoscaler1_beta2 as autoscaler1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use autoscaler1_beta2::AutoscalerHub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AutoscalerHub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)` and `list(...)`
// to build up your call.
let rb = hub.zone_operations();

Methods

impl<'a, C, A> ZoneOperationMethods<'a, C, A>

fn list(&self, project: &str, zone: &str) -> ZoneOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of operation resources contained within the specified zone.

Arguments

project - No description provided.

zone - No description provided.

fn delete(&self, project: &str, zone: &str, operation: &str) -> ZoneOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified zone-specific operation resource.

Arguments

project - No description provided.

zone - No description provided.

operation - No description provided.

fn get(&self, project: &str, zone: &str, operation: &str) -> ZoneOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified zone-specific operation resource.

Arguments

project - No description provided.

zone - No description provided.

operation - No description provided.